requirements.txt
pandas>=2.2
numpy>=1.26
pyarrow>=15.0
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
fiona>=1.9
requests>=2.31
tenacity>=8.2
python-dotenv>=1.0
PyYAML>=6.0
notion-client>=2.2
structlog>=24.1
pytest>=8.0
